添加雪花算法

+++ model/User.go
@@ -6,7 +6,7 @@ import (
 )
 
 type User struct {
-	Id        uint   `gorm:"primarykey"`
+	Id        int64  `gorm:"primarykey"`
 	Phone     string `gorm:"column:phone"`
 	Password  string `gorm:"column:password"`
 	Username  string `gorm:"column:username"`
